type2/pyproject.toml
[tool.poetry]
name = "hyrda_config_mvp_type02"
version = "0.1.0"
description = ""
authors = ["gatheluck <gatheluck+tech@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
hydra-core = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"